Skills Catalog!
Did some awesome stuff... not sure if this is the best solution but it
works. Definitely need a code review. I kept running into an issue where
the autocomplete could not hand my array complex object so I had to do a
work around. Either way this resolves issue #9. :hammer: